requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
redis[hiredis]==5.1.0
httpx==0.27.0
pydantic-settings==2.5.0
cryptography==43.0.0
python-dotenv==1.0.1
pgvector==0.3.0
python-jose[cryptography]==3.3.0
alembic==1.13.0
pytest==8.3.0
pytest-asyncio==0.24.0
hypothesis==6.112.0
respx==0.21.0
